refactor/selene_api
