+++ .github/workflows/unit_tests.yml
@@ -55,6 +55,7 @@ jobs:
         run: |
           pip install -r requirements/tests.txt
           pip install ./test/unittests/common_query/ovos_tskill_fakewiki
+          pip install git+https://github.com/OpenVoiceOS/selene_api
       - name: Run unittests
         run: |
           pytest --cov=mycroft --cov-report xml test/unittests
