+++ mycroft/configuration/mycroft.conf
@@ -121,8 +121,22 @@
     // relative to "data_dir"
     "directory": "skills",
 
-    // used by selene to display skill settings in web interface
-    "upload_skill_manifest": true,
+    // used by selene for marketplace integration in web interface
+    // NOTE: selene seems to work fine without this data
+    // no need to advertise installed skills if they dont have settings
+    // this data was tightly coupled to msm and is now mostly useless for selene
+    "upload_skill_manifest": false,
+
+    // if "sync_skill_settings" is enabled in "server" section
+    // should skill settingg changes on device be upload to selene?
+    // NOTE: this should be true, but it was removed (bug ?) in selene at some point
+    // old endpoints are however still available so functionality has been restored here
+    // it is only configurable in case you want to retain "old" mycroft-core behaviour
+    // or in case selene deprecates the old endpoint
+    "sync2way": true,
+    // values in skill settings missing in settingsmeta wont show up in selene
+    // this flag auto generates settingsmeta on the fly
+    "autogen_meta": true,
 
     // blacklisted skills to not load
     // NB: This is the basename() of the directory where the skill lives, so if
