+++ mycroft/deprecated/skills/msm_wrapper.py
