+++ mycroft/deprecated/stt.py
@@ -2,3 +2,7 @@
 from mycroft.listener.stt import *
 from ovos_plugin_manager.templates.stt import STT, TokenSTT, GoogleJsonSTT, \
     StreamingSTT, StreamThread, BasicSTT, KeySTT
+
+# for compat in case its being imported elsewhere
+from ovos_stt_plugin_selene import SeleneSTT as MycroftSTT
+
