+++ requirements/requirements.txt
@@ -8,6 +8,7 @@ combo-lock~=0.2
 PyYAML~=5.4
 watchdog
 
+selene_api~=0.0.2
 ovos-config~=0.0.1, >=0.0.4
 ovos-utils~=0.0, >=0.0.23
 ovos-plugin-manager~=0.0, >=0.0.18
