+++ test/unittests/api/test_api.py
@@ -14,13 +14,16 @@
 #
 import unittest
 from copy import copy
-
+import selene_api
 from unittest.mock import MagicMock, patch
 
 import mycroft.api
 import mycroft.configuration
 from mycroft.configuration import Configuration
 from test.util import base_config
+
+# TODO move test to selene_api repo
+
 CONFIG = base_config()
 CONFIG.merge(
     {
@@ -34,8 +37,7 @@ CONFIG.merge(
     }
 )
 
-
-mycroft.api.requests.post = MagicMock()
+selene_api.api.requests.post = MagicMock()
 
 
 def create_identity(uuid, expired=False):
@@ -56,7 +58,7 @@ def create_response(status, json=None, url='', data=''):
 
 @patch.dict(Configuration._Configuration__patch, CONFIG)
 class TestApi(unittest.TestCase):
-    @patch('mycroft.api.IdentityManager.get')
+    @patch('selene_api.identity.IdentityManager.get')
     def test_init(self, mock_identity_get):
         mock_identity_get.return_value = create_identity('1234')
         a = mycroft.api.Api('test-path')
@@ -65,8 +67,8 @@ class TestApi(unittest.TestCase):
         self.assertEqual(a.identity.uuid, '1234')
 
     @unittest.skip("requires backend to be enabled, TODO refactor test!")
-    @patch('mycroft.api.IdentityManager')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager')
+    @patch('selene_api.api.requests.request')
     def test_send(self, mock_request, mock_identity_manager):
         # Setup an OK response
         mock_response_ok = create_response(200, {})
@@ -83,13 +85,13 @@ class TestApi(unittest.TestCase):
 
         # check that a 300+ status code generates Exception
         mock_request.return_value = mock_response_301
-        with self.assertRaises(mycroft.api.HTTPError):
+        with self.assertRaises(selene_api.exceptions.HTTPError):
             a.send(req)
 
         # Check 401
         mock_request.return_value = mock_response_401
         req = {'path': '', 'headers': {}}
-        with self.assertRaises(mycroft.api.HTTPError):
+        with self.assertRaises(selene_api.exceptions.HTTPError):
             a.send(req)
 
         # Check refresh token
@@ -98,79 +100,80 @@ class TestApi(unittest.TestCase):
                                     mock_response_ok]
         req = {'path': 'something', 'headers': {}}
         a.send(req)
-        self.assertTrue(mycroft.api.IdentityManager.save.called)
+        self.assertTrue(selene_api.identity.IdentityManager.save.called)
 
 
-@unittest.skip("requires backend to be enabled, TODO refactor test!")
+#@unittest.skip("requires backend to be enabled, TODO refactor test!")
 @patch.dict(Configuration._Configuration__patch, CONFIG)
 class TestDeviceApi(unittest.TestCase):
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.request')
     def test_init(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity_get.return_value = create_identity('1234')
 
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         self.assertEqual(device.identity.uuid, '1234')
-        self.assertEqual(device.path, 'device')
+        self.assertTrue(device.url.endswith("/device"))
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.post')
     def test_device_activate(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity_get.return_value = create_identity('1234')
         # Test activate
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.activate('state', 'token')
         json = mock_request.call_args[1]['json']
         self.assertEqual(json['state'], 'state')
         self.assertEqual(json['token'], 'token')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity_get.return_value = create_identity('1234')
         # Test get
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get()
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(url, 'https://api-test.mycroft.ai/v1/device/1234')
 
-    @patch('mycroft.api.IdentityManager.update')
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.update')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get_code(self, mock_request, mock_identity_get,
                              mock_identit_update):
         mock_request.return_value = create_response(200, '123ABC')
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         ret = device.get_code('state')
         self.assertEqual(ret, '123ABC')
-        url = mock_request.call_args[0][1]
-        self.assertEqual(
-            url, 'https://api-test.mycroft.ai/v1/device/code?state=state')
+        url = mock_request.call_args[0][0]
+        params = mock_request.call_args[1]
+        self.assertEqual(url, 'https://api-test.mycroft.ai/v1/device/code')
+        self.assertEqual(params["params"], {"state": "state"})
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get_settings(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get_settings()
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/setting')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.post')
     def test_device_report_metric(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.report_metric('mymetric', {'data': 'mydata'})
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         params = mock_request.call_args[1]
 
         content_type = params['headers']['Content-Type']
@@ -180,14 +183,14 @@ class TestDeviceApi(unittest.TestCase):
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/metric/mymetric')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.put')
     def test_device_send_email(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.send_email('title', 'body', 'sender')
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         params = mock_request.call_args[1]
 
         content_type = params['headers']['Content-Type']
@@ -197,37 +200,37 @@ class TestDeviceApi(unittest.TestCase):
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/message')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get_oauth_token(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get_oauth_token(1)
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
 
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/token/1')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get_location(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get_location()
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/location')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_device_get_subscription(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get_subscription()
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/subscription')
 
@@ -240,14 +243,14 @@ class TestDeviceApi(unittest.TestCase):
         mock_request.return_value = create_response(200, {'@type': 'yearly'})
         self.assertTrue(device.is_subscriber)
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.put')
     def test_device_upload_skills_data(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.upload_skills_data({})
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         data = mock_request.call_args[1]['json']
 
         # Check that the correct url is called
@@ -261,53 +264,53 @@ class TestDeviceApi(unittest.TestCase):
         with self.assertRaises(ValueError):
             device.upload_skills_data('This isn\'t right at all')
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_stt(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        stt = mycroft.api.STTApi('stt')
-        self.assertEqual(stt.path, 'stt')
+        stt = selene_api.api.STTApi('stt')
+        self.assertTrue(stt.url.endswith('stt'))
 
-    @patch('mycroft.api.IdentityManager.get')
-    @patch('mycroft.api.requests.request')
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.post')
     def test_stt_stt(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        stt = mycroft.api.STTApi('stt')
+        stt = selene_api.api.STTApi('https://api-test.mycroft.ai')
         stt.stt('La la la', 'en-US', 1)
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(url, 'https://api-test.mycroft.ai/v1/stt')
         data = mock_request.call_args[1].get('data')
         self.assertEqual(data, 'La la la')
         params = mock_request.call_args[1].get('params')
         self.assertEqual(params['lang'], 'en-US')
 
-    @patch('mycroft.api.IdentityManager.load')
+    @patch('selene_api.identity.IdentityManager.load')
     def test_has_been_paired(self, mock_identity_load):
         # reset pairing cache
         mock_identity = MagicMock()
         mock_identity_load.return_value = mock_identity
         # Test None
         mock_identity.uuid = None
-        self.assertFalse(mycroft.api.has_been_paired())
+        self.assertFalse(selene_api.pairing.has_been_paired())
         # Test empty string
         mock_identity.uuid = ""
-        self.assertFalse(mycroft.api.has_been_paired())
+        self.assertFalse(selene_api.pairing.has_been_paired())
         # Test actual id number
         mock_identity.uuid = "1234"
-        self.assertTrue(mycroft.api.has_been_paired())
+        self.assertTrue(selene_api.pairing.has_been_paired())
 
 
-@unittest.skip("requires backend to be enabled, TODO refactor test!")
 @patch.dict(Configuration._Configuration__patch, CONFIG)
-@patch('mycroft.api.IdentityManager.get')
-@patch('mycroft.api.requests.request')
 class TestSettingsMeta(unittest.TestCase):
+
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.put')
     def test_upload_meta(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
 
         settings_meta = {
             'name': 'TestMeta',
@@ -328,37 +331,34 @@ class TestSettingsMeta(unittest.TestCase):
             }
         }
         device.upload_skill_metadata(settings_meta)
-        url = mock_request.call_args[0][1]
-        method = mock_request.call_args[0][0]
+        url = mock_request.call_args[0][0]
         params = mock_request.call_args[1]
 
         content_type = params['headers']['Content-Type']
         self.assertEqual(content_type, 'application/json')
-        self.assertEqual(method, 'PUT')
         self.assertEqual(params['json'], settings_meta)
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/settingsMeta')
 
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_get_skill_settings(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200, {})
         mock_identity_get.return_value = create_identity('1234')
-        device = mycroft.api.DeviceApi()
+        device = selene_api.api.DeviceApi()
         device.get_skill_settings()
-        method = mock_request.call_args[0][0]
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         params = mock_request.call_args[1]
 
-        self.assertEqual(method, 'GET')
         self.assertEqual(
             url, 'https://api-test.mycroft.ai/v1/device/1234/skill/settings')
 
 
-@unittest.skip("requires backend to be enabled, TODO refactor test!")
 @patch.dict(Configuration._Configuration__patch, CONFIG)
-@patch('mycroft.api._paired_cache', False)
-@patch('mycroft.api.IdentityManager.get')
-@patch('mycroft.api.requests.request')
+@patch('selene_api.pairing._paired_cache', False)
 class TestIsPaired(unittest.TestCase):
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_is_paired_true(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity = MagicMock()
@@ -368,12 +368,14 @@ class TestIsPaired(unittest.TestCase):
         num_calls = mock_identity_get.num_calls
         # reset paired cache
 
-        self.assertTrue(mycroft.api.is_paired())
+        self.assertTrue(selene_api.pairing.is_paired())
 
         self.assertEqual(num_calls, mock_identity_get.num_calls)
-        url = mock_request.call_args[0][1]
+        url = mock_request.call_args[0][0]
         self.assertEqual(url, 'https://api-test.mycroft.ai/v1/device/1234')
 
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_is_paired_false_local(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(200)
         mock_identity = MagicMock()
@@ -381,19 +383,25 @@ class TestIsPaired(unittest.TestCase):
         mock_identity.uuid = ''
         mock_identity_get.return_value = mock_identity
 
-        self.assertFalse(mycroft.api.is_paired())
+        self.assertFalse(selene_api.pairing.is_paired())
         mock_identity.uuid = None
-        self.assertFalse(mycroft.api.is_paired())
+        self.assertFalse(selene_api.pairing.is_paired())
 
+    @unittest.skip("TODO - refactor/fix test")
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_is_paired_false_remote(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(401)
         mock_identity = MagicMock()
         mock_identity.is_expired.return_value = False
         mock_identity.uuid = '1234'
         mock_identity_get.return_value = mock_identity
+        selene_api.pairing._paired_cache = False
+        self.assertFalse(selene_api.pairing.is_paired())
 
-        self.assertFalse(mycroft.api.is_paired())
-
+    @unittest.skip("TODO - refactor/fix test")
+    @patch('selene_api.identity.IdentityManager.get')
+    @patch('selene_api.api.requests.get')
     def test_is_paired_error_remote(self, mock_request, mock_identity_get):
         mock_request.return_value = create_response(500)
         mock_identity = MagicMock()
@@ -401,7 +409,7 @@ class TestIsPaired(unittest.TestCase):
         mock_identity.uuid = '1234'
         mock_identity_get.return_value = mock_identity
 
-        self.assertFalse(mycroft.api.is_paired())
+        self.assertFalse(selene_api.pairing.is_paired())
 
-        with self.assertRaises(mycroft.api.BackendDown):
-            mycroft.api.is_paired(ignore_errors=False)
+        with self.assertRaises(selene_api.exceptions.BackendDown):
+            selene_api.pairing.is_paired(ignore_errors=False)
